ai/requirements.txt
django==4.2.5
pandas==2.1.1
numpy==1.26.0
scikit-learn==1.3.1
Pillow==10.1.0
torch==2.1.0
torchvision==0.16.0
djangorestframework==3.14.0
